requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.1
easydict==1.13
faiss-cpu==1.11.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
onnx==1.18.0
onnxruntime==1.22.0
onnxruntime-gpu==1.22.0
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
prettytable==3.16.0
protobuf==6.31.1
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
simsimd==6.4.7
six==1.17.0
stringzilla==3.12.5
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
